feat: add GlobalExceptionHandler with error responses

+++ backend/src/main/java/br/com/dinacare/repository/procedure/ProcedureRepository.java
@@ -4,6 +4,7 @@ import br.com.dinacare.domain.procedure.Procedure;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -11,5 +12,5 @@ import java.util.UUID;
 public interface ProcedureRepository extends JpaRepository<Procedure, UUID> {
 
     List<Procedure> findByActiveTrue();
-    List<Procedure> findByPriceBetween(Double min, Double max);
+    List<Procedure> findByPriceBetween(BigDecimal min, BigDecimal max);
 }
